refactor: Document library use of cargo2port
When src/lib.rs was added, cargo2port was made available for library
use, but the functions were still organized in such a way as to reflect
the cargo2port command line.

This refactors the library code to make the functions more usable
outside of the cargo2port command line context.